check for auth

+++ frontend/src/components/PostList.jsx
@@ -6,9 +6,12 @@ function PostList({ post }) {
   const { dispatch } = usePostContext();
 
   const handleDelete = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(`http://localhost:4000/api/post/${post._id}`, {
       method: "DELETE",
-      headers: { Authentication: `Bearer ${user.token}` },
+      headers: { Authorization: `Bearer ${user.token}` },
     });
 
     const json = await response.json();
@@ -19,6 +22,7 @@ function PostList({ post }) {
   };
   return (
     <div className="border mx-20 p-6 my-6 flex justify-between">
+      <p>{user?.name}</p>
       <div>{post?.post}</div>
       <button onClick={handleDelete}>Delete</button>
     </div>
